+++ contracts/test/strategies/base/TestRepayStrategy.sol
@@ -187,7 +187,7 @@ contract TestRepayStrategy is RepayStrategy, BorrowStrategy {
     }
 
     function updateLoan(LibStorage.Loan storage _loan) internal override returns(uint256){
-        uint96 rateIndex = borrowRate;
+        uint80 rateIndex = borrowRate;
         s.accFeeIndex = rateIndex;
         return updateLoanLiquidity(_loan, rateIndex);
     }
@@ -243,7 +243,7 @@ contract TestRepayStrategy is RepayStrategy, BorrowStrategy {
     }
 
     function getPoolData() external virtual view returns(uint256 LP_TOKEN_BALANCE, uint256 LP_TOKEN_BORROWED, uint40 LAST_BLOCK_NUMBER,
-        uint96 accFeeIndex, uint256 LP_TOKEN_BORROWED_PLUS_INTEREST, uint128 LP_INVARIANT, uint128 BORROWED_INVARIANT, uint128[] memory CFMM_RESERVES) {
+        uint80 accFeeIndex, uint256 LP_TOKEN_BORROWED_PLUS_INTEREST, uint128 LP_INVARIANT, uint128 BORROWED_INVARIANT, uint128[] memory CFMM_RESERVES) {
         LP_TOKEN_BALANCE = s.LP_TOKEN_BALANCE;
         LP_TOKEN_BORROWED = s.LP_TOKEN_BORROWED;
         LAST_BLOCK_NUMBER = s.LAST_BLOCK_NUMBER;
@@ -269,7 +269,7 @@ contract TestRepayStrategy is RepayStrategy, BorrowStrategy {
         updateIndex();
     }
 
-    function setAccFeeIndex(uint96 accFeeIndex) public virtual {
+    function setAccFeeIndex(uint80 accFeeIndex) public virtual {
         s.accFeeIndex = accFeeIndex;
     }
